requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.0
torchvision==0.16.0
efficientnet-pytorch==0.7.1
onnxruntime==1.16.3
pillow==10.0.1
python-multipart==0.0.6
numpy==1.24.3
